+++ code/src/app/components/pages/landing.tsx
@@ -1,6 +1,7 @@
 export default function Landing() {
     return (
         <div>
+            <h1>Landing</h1>
             <p>
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
                 Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
