+++ code/src/app/components/pages/offerings.tsx
@@ -1,7 +1,7 @@
 export default function Offerings() {
     return (
         <div>
-
+            <h1>Offerings</h1>
             <p>
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
                 Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
