+++ code/src/app/constants/section.tsx
@@ -1,25 +1,21 @@
 import { SectionContent } from "../models/sectionContent";
-import Landing from "../components/sections/landing";
-import Offerings from "../components/sections/offerings";
-import Domains from "../components/sections/domains";
+import Landing from "../components/pages/landing";
+import Offerings from "../components/pages/offerings";
+import Domains from "../components/pages/domains";
 export const sections: SectionContent[] = [
     {
         id: 'landing',
-        title: "Landing Section",
         mainHeader: true,
         content: <Landing />
-
     },
     {
         id: "domains",
-        title: "Domain Section",
         mainHeader: false,
         content: <Domains />
 
     },
     {
         id: "offerings",
-        title: "OfferingSection",
         mainHeader: false,
         content: <Offerings />
