+++ code/src/app/models/sectionContent.ts
@@ -3,7 +3,6 @@ import { type ReactElement } from "react";
 
 export interface SectionContent {
   id: string;
-  title: string;
   mainHeader?: boolean;
   content: ReactElement;
 }
